Extract workspace command plan executor
