+++ Sources/QuillCodeApp/WorkspaceModel.swift
@@ -1221,55 +1221,7 @@ public final class QuillCodeWorkspaceModel {
     }
 
     @discardableResult
-    public func runWorkspaceCommand(_ commandID: String, workspaceRoot: URL) -> Bool {
-        guard let plan = WorkspaceCommandPlan(commandID: commandID) else { return false }
-        switch plan {
-        case .localEnvironmentAction(let actionID):
-            return runLocalEnvironmentAction(actionID, workspaceRoot: workspaceRoot)
-        case .deleteMemory(let id):
-            return deleteGlobalMemory(id: id)
-        case .updateAutomationStatus(let id, let status):
-            return updateAutomationStatus(id: id, status: status)
-        case .runAutomation(let id):
-            return runAutomation(id: id) != nil
-        case .deleteAutomation(let id):
-            return deleteAutomation(id: id)
-        case .createThreadFollowUpAfter(let seconds):
-            return createThreadFollowUpAutomation(after: seconds) != nil
-        case .createWorkspaceScheduleAfter(let seconds):
-            return createWorkspaceScheduleAutomation(after: seconds) != nil
-        case .createThreadFollowUpEvery(let recurrence):
-            return createThreadFollowUpAutomation(every: recurrence) != nil
-        case .createWorkspaceScheduleEvery(let recurrence):
-            return createWorkspaceScheduleAutomation(every: recurrence) != nil
-        case .startMCPServer(let id):
-            return startMCPServer(id: id, workspaceRoot: workspaceRoot)
-        case .stopMCPServer(let id):
-            return stopMCPServer(id: id)
-        case .updateExtension(let id):
-            return runProjectExtensionUpdate(id: id, workspaceRoot: workspaceRoot)
-        case .toggleThreadSelection(let id):
-            toggleSidebarThreadSelection(id)
-            return true
-        case .toggleActivitySection(let section):
-            toggleActivitySection(section)
-            return true
-        case .setDraft(let draft):
-            setDraft(draft)
-            return true
-        case .runTool(let toolName):
-            runToolCall(
-                ToolCall(name: toolName, argumentsJSON: "{}"),
-                workspaceRoot: workspaceRoot
-            )
-            return true
-        case .action(let action):
-            return runWorkspaceCommandAction(action)
-        }
-    }
-
-    @discardableResult
-    private func startMCPServer(id: String, workspaceRoot: URL) -> Bool {
+    func startMCPServer(id: String, workspaceRoot: URL) -> Bool {
         guard let manifest = selectedProject?.extensionManifests.first(where: {
             $0.id == id && $0.kind == .mcpServer
         }) else {
@@ -1294,7 +1246,7 @@ public final class QuillCodeWorkspaceModel {
     }
 
     @discardableResult
-    private func stopMCPServer(id: String) -> Bool {
+    func stopMCPServer(id: String) -> Bool {
         guard let manifest = selectedProject?.extensionManifests.first(where: {
             $0.id == id && $0.kind == .mcpServer
         }) else {
@@ -1334,7 +1286,7 @@ public final class QuillCodeWorkspaceModel {
     }
 
     @discardableResult
-    private func deleteGlobalMemory(id: String) -> Bool {
+    func deleteGlobalMemory(id: String) -> Bool {
         guard let mutation = WorkspaceMemoryEngine.deleteGlobal(id: id, directory: globalMemoryDirectory) else {
             return false
         }
